+++ packages/grafana-runtime/src/index.ts
@@ -3,3 +3,5 @@ export * from './config';
 export * from './types';
 export { loadPluginCss, SystemJS } from './utils/plugin';
 export { reportMetaAnalytics } from './utils/analytics';
+
+export { DataSourceWithBackend } from './utils/DataSourceWithBackend';
